Rename time.h to sweep_time.h.
It's easy to have compiler errors with it named time.h, as this
is the same name as a system header.
