+++ libsweep/time_unix.c
@@ -1,6 +1,6 @@
 #define _POSIX_C_SOURCE 200809L
 
-#include "time.h"
+#include "sweep_time.h"
 #include "sweep.h"
 
 #include <errno.h>
